Fixed some compilation errors.
Rev: src/Makefile.in:1.3
Rev: src/bignum.c:1.2
Rev: src/bignum.h:1.2
Rev: src/debug.c:1.3
Rev: src/debug.h:1.2
Rev: src/decrypt.c:1.3
Rev: src/decrypt.h:1.2
Rev: src/encrypt.c:1.3
Rev: src/encrypt.h:1.2
Rev: src/format.c:1.3
Rev: src/format.h:1.2
Rev: src/packet_dispatch.c:1.2
Rev: src/packet_dispatch.h:1.2
Rev: src/pad.h:1.2
Rev: src/parse.c:1.3
Rev: src/parse.h:1.3
Rev: src/process_atoms:1.3
Rev: src/transport.h:1.4
Rev: src/unpad.h:1.2

+++ src/format.h
@@ -8,6 +8,10 @@
 
 #include <stdarg.h>
 
+#include "atoms.h"
+#include "bignum.h"
+#include "transport.h"
+
 /* Format strings can contain the following %-specifications:
  *
  * %%  Insert a %-sign
@@ -31,7 +35,7 @@
  */
 
 UINT32 ssh_vformat_length(char *format, va_list args);
-void ssh_vformat(char *format, UINT *buffer, va_list args);
+void ssh_vformat(char *format, UINT8 *buffer, va_list args);
 struct simple_packet *ssh_format(char *format, ...);
 
 #endif /* LSH_FORMAT_H_INCLUDED */
